event-intelligence-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
yfinance==0.2.36
newsapi-python==0.2.7
pydantic==2.5.3
python-dotenv==1.0.0
